small improvements

+++ server/src/main/kotlin/com/unde/server/configuration/adb/AdbManager.kt
@@ -31,16 +31,12 @@ object AdbManager {
      */
     internal fun setup(application: Application) {
         adbDevicesJob?.cancel()
-        adbDevicesJob = application.launch(Dispatchers.Default) {
+        adbDevicesJob = application.launch(Dispatchers.IO) {
             while (isActive) {
-                val devicesDeferred = application.async(Dispatchers.Default) { getConnectedDevices() }
+                val connectedDevicesDeferred = application.async { getConnectedDevices() }
+                val reversedDevicesDeferred = application.async { getReversedDevices() }
                 ensureActive()
-                with(devicesDeferred.await()) {
-                    if (isNotEmpty()) {
-                        ensureActive()
-                        setupAdbReverse(this)
-                    }
-                }
+                setupAdbReverse(connectedDevicesDeferred.await(), reversedDevicesDeferred.await())
                 delay(Time.DEFAULT_ADB_TASK_DELAY)
             }
         }
@@ -63,35 +59,62 @@ object AdbManager {
         val devices = process.inputStream.bufferedReader().useLines { lines ->
             lines.drop(1) // Skip the header "List of devices attached"
                 .filter { it.contains("device") && !it.contains("unauthorized") }
-                .map { line ->
-                    return@map line.split(" ").firstOrNull { it.isNotBlank() }.orEmpty()
-                }.filter { it.isNotBlank() }.toList()
+                .map { line -> line.split(" ").firstOrNull { it.isNotBlank() }.orEmpty() }
+                .filter { it.isNotBlank() }.toList()
         }
         currentCoroutineContext().ensureActive()
         LOGGER.info("Adb devices: $devices")
         process.waitFor()
+        process.destroy()
+        process.destroyForcibly()
         return devices
     }
 
-    private suspend fun setupAdbReverse(devices: List<String> = emptyList()) {
+    private suspend fun getReversedDevices(): List<String> {
+        currentCoroutineContext().ensureActive()
+        LOGGER.info("Trying to get reversed devices")
+        val process = ProcessBuilder("adb", "reverse", "--list").redirectErrorStream(true).start()
+        currentCoroutineContext().ensureActive()
+        val devices = process.inputStream.bufferedReader().useLines { lines ->
+            lines
+                .filter { !it.contains("error:") }
+                .map { line -> line.split(" ").firstOrNull { it.isNotBlank() }.orEmpty() }
+                .filter { it.isNotBlank() }.toList()
+        }
+        currentCoroutineContext().ensureActive()
+        LOGGER.info("ADB REVERSED DEVICES: $devices")
+        process.waitFor()
+        process.destroy()
+        process.destroyForcibly()
+        return devices
+    }
+
+    private suspend fun setupAdbReverse(availableDevices: List<String> = emptyList(), reversedDevices: List<String>) {
+        val connectedDevices = availableDevices.associateWith { reversedDevices.contains(it) }.toMutableMap()
         try {
-            devices.forEach {
-                currentCoroutineContext().ensureActive()
-                val process = ProcessBuilder("adb", "-s", it, "reverse", "tcp:8081", "tcp:8081")
-                    .redirectErrorStream(true)
-                    .start()
-                LOGGER.info("Trying to setup adb reverse for device: $it")
-                currentCoroutineContext().ensureActive()
-                val result = process.inputStream.bufferedReader().use(BufferedReader::readText)
-                currentCoroutineContext().ensureActive()
-                if (adbErrors.contains(result)) {
-                    LOGGER.warn("ADB reverse was not installed because of [$result]")
-                    throw IllegalStateException("ADB reverse was not installed because of [$result]")
+            connectedDevices.forEach { deviceToReverse ->
+                if (!deviceToReverse.value) {
+                    currentCoroutineContext().ensureActive()
+                    val process = ProcessBuilder("adb", "-s", deviceToReverse.key, "reverse", "tcp:8081", "tcp:8081")
+                        .redirectErrorStream(true)
+                        .start()
+                    LOGGER.info("Trying to setup adb reverse for device: ${deviceToReverse.key}")
+                    currentCoroutineContext().ensureActive()
+                    val result = process.inputStream.bufferedReader().use(BufferedReader::readText)
+                    currentCoroutineContext().ensureActive()
+                    if (adbErrors.contains(result)) {
+                        LOGGER.warn("ADB reverse was not installed because of [$result]")
+                        throw IllegalStateException("ADB reverse was not installed because of [$result]")
+                    }
+                    // Wait for the process to finish
+                    val exitCode = process.waitFor()
+                    if (exitCode == 0) {
+                        connectedDevices[deviceToReverse.key] = true
+                    }
+                    process.destroy()
+                    process.destroyForcibly()
+                    LOGGER.info("ADB REVERSE: \n| result: $result \n| exitCode: $exitCode")
                 }
-                // Wait for the process to finish
-                val exitCode = process.waitFor()
-                process.destroyForcibly()
-                LOGGER.info("ADB Reverse result: $result | exitCode: $exitCode")
             }
         } catch (e: Exception) {
             LOGGER.warn("ADB reverse cannot be configured because of [${e.stackTrace}]")
